+++ spec/dop_common/step_spec.rb
@@ -17,57 +17,33 @@ describe DopCommon::Step do
     end
   end
 
-  describe '#nodes' do
-    it 'returns an array with nodes if correctly specified' do
-      step = DopCommon::Step.new({:name => 'foo', :nodes => ['foo']})
-      expect(step.nodes).to eq ['foo']
-    end
-    it 'returns an empty array if the key is missing' do
-      step = DopCommon::Step.new({:name => 'foo'})
-      expect(step.nodes).to eq []
-    end
-    it 'returns an array with one element if the value is a string' do
-      step = DopCommon::Step.new({:name => 'foo', :nodes => 'foo'})
-      expect(step.nodes).to eq ['foo']
-    end
-    it 'throws an exception if the value is something other than an array or a string' do
-      step = DopCommon::Step.new({:name => 'foo', :nodes => 1})
-      expect{step.nodes}.to raise_error DopCommon::PlanParsingError
-    end
-    it 'throws an exception if the array contains something other than a string' do
-      step = DopCommon::Step.new({:name => 'foo', :nodes => ['foo', 1]})
-      expect{step.nodes}.to raise_error DopCommon::PlanParsingError
-    end
-    it 'throws an exception if the array contains an invalid regexp' do
-      step = DopCommon::Step.new({:name => 'foo', :nodes => ['/][/']})
-      expect{step.nodes}.to raise_error DopCommon::PlanParsingError
-    end
-  end
 
-  describe '#roles' do
-    it 'returns an array with roles if correctly specified' do
-      step = DopCommon::Step.new({:name => 'foo', :roles => ['foo']})
-      expect(step.roles).to eq ['foo']
-    end
-    it 'returns an empty array if the key is missing' do
-      step = DopCommon::Step.new({:name => 'foo'})
-      expect(step.roles).to eq []
-    end
-    it 'returns an array with one element if the value is a string' do
-      step = DopCommon::Step.new({:name => 'foo', :roles => 'foo'})
-      expect(step.roles).to eq ['foo']
-    end
-    it 'throws an exception if the value is something other than an array or a string' do
-      step = DopCommon::Step.new({:name => 'foo', :roles => 1})
-      expect{step.roles}.to raise_error DopCommon::PlanParsingError
-    end
-    it 'throws an exception if the array contains something other than a string' do
-      step = DopCommon::Step.new({:name => 'foo', :roles => ['foo', 1]})
-      expect{step.roles}.to raise_error DopCommon::PlanParsingError
-    end
-    it 'throws an exception if the array contains an invalid regexp' do
-      step = DopCommon::Step.new({:name => 'foo', :roles => ['/][/']})
-      expect{step.roles}.to raise_error DopCommon::PlanParsingError
+  [:nodes, :exclude_nodes, :roles, :exclude_roles].each do |method_name|
+    describe '#' + method_name.to_s do
+      it 'returns an array with entries if correctly specified' do
+        step = DopCommon::Step.new({:name => 'foo', method_name => ['foo']})
+        expect(step.send(method_name)).to eq(['foo'])
+      end
+      it 'returns an empty array if the key is missing' do
+        step = DopCommon::Step.new({:name => 'foo'})
+        expect(step.send(method_name)).to eq([])
+      end
+      it 'returns an array with one element if the value is a string' do
+        step = DopCommon::Step.new({:name => 'foo', method_name => 'foo'})
+        expect(step.send(method_name)).to eq(['foo'])
+      end
+      it 'throws an exception if the value is something other than an array or a string' do
+        step = DopCommon::Step.new({:name => 'foo', method_name => 1})
+        expect{step.send(method_name)}.to raise_error DopCommon::PlanParsingError
+      end
+      it 'throws an exception if the array contains something other than a string' do
+        step = DopCommon::Step.new({:name => 'foo', method_name => ['foo', 1]})
+        expect{step.send(method_name)}.to raise_error DopCommon::PlanParsingError
+      end
+      it 'throws an exception if the array contains an invalid regexp' do
+        step = DopCommon::Step.new({:name => 'foo', method_name => ['/][/']})
+        expect{step.send(method_name)}.to raise_error DopCommon::PlanParsingError
+      end
     end
   end
 
@@ -75,7 +51,7 @@ describe DopCommon::Step do
     describe '#' + method_name.to_s do
       it 'returns a Hash if correctly specified' do
         step = DopCommon::Step.new({:name => 'foo', method_name => {'my_alias' => 'foo'}})
-        expect(step.send(method_name)).to eq({'my_alias' => 'foo'})
+        expect(step.send(method_name)).to eq({'my_alias' => ['foo']})
       end
       it 'returns an empty hash if the key is missing' do
         step = DopCommon::Step.new({:name => 'foo'})